Remove eventsource request from SearchContexts call
